requirements.txt
# Runtime dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
pyyaml>=6.0.1
tenacity>=8.2.0
tqdm>=4.66.0

# Parser dependencies
tabula-py>=2.9.0
camelot-py[cv]>=0.11.0
pdfplumber>=0.10.0
pandas>=2.1.0
openpyxl>=3.1.0
Flask>=2.3.0

# Optional OCR (install separately if needed)
# pytesseract>=0.3.10
# pdf2image>=1.17.0

# Dev / test
pytest>=7.4.0
pytest-cov>=4.1.0
reportlab>=4.0.0
